refactor: 행사 상세 페이지 로직을 커스텀 훅으로 분리
