+++ src/app/event/[eventId]/components/event-form/EventForm.tsx
@@ -1,54 +1,47 @@
 'use client';
 
-import HubButton from './HubButton';
-import DateButton from './DateButton';
-import BottomBar from './BottomBar';
+import HubButton from '../HubButton';
+import DateButton from '../DateButton';
+import BottomBar from '../BottomBar';
 import useBottomSheet from '@/hooks/useBottomSheet';
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useMemo, useState } from 'react';
 import BottomSheet from '@/components/bottom-sheet/BottomSheet';
-import CommonDateStep from './steps/common-steps/CommonDateStep';
-import CommonSidoStep from './steps/common-steps/CommonSidoStep';
-import ReservationHubsStep from './steps/reservation-steps/ReservationHubsStep';
-import ReservationTripTypeStep from './steps/reservation-steps/ReservationTripTypeStep';
-import ReservationInfoStep from './steps/reservation-steps/ReservationInfoStep';
-import DemandHubInfoStep from './steps/demand-steps/DemandHubInfoStep';
-import ExtraSidoInfoStep from './steps/extra-steps/ExtraSidoInfoStep';
-import ExtraOpenSidoStep from './steps/extra-steps/ExtraOpenSidoStep';
-import DemandHubsStep from './steps/demand-steps/DemandHubsStep';
-import ExtraDuplicateHubStep from './steps/extra-steps/ExtraDuplicateHubStep';
-import ExtraSeatAlarmStep from './steps/extra-steps/ExtraSeatAlarmStep';
-import DemandTripTypeStep from './steps/demand-steps/DemandTripTypeStep';
+import CommonDateStep from '../steps/common-steps/CommonDateStep';
+import CommonSidoStep from '../steps/common-steps/CommonSidoStep';
+import ReservationHubsStep from '../steps/reservation-steps/ReservationHubsStep';
+import ReservationTripTypeStep from '../steps/reservation-steps/ReservationTripTypeStep';
+import ReservationInfoStep from '../steps/reservation-steps/ReservationInfoStep';
+import DemandHubInfoStep from '../steps/demand-steps/DemandHubInfoStep';
+import ExtraSidoInfoStep from '../steps/extra-steps/ExtraSidoInfoStep';
+import ExtraOpenSidoStep from '../steps/extra-steps/ExtraOpenSidoStep';
+import DemandHubsStep from '../steps/demand-steps/DemandHubsStep';
+import ExtraDuplicateHubStep from '../steps/extra-steps/ExtraDuplicateHubStep';
+import ExtraSeatAlarmStep from '../steps/extra-steps/ExtraSeatAlarmStep';
+import DemandTripTypeStep from '../steps/demand-steps/DemandTripTypeStep';
 import useFunnel from '@/hooks/useFunnel';
-import {
-  EVENT_FORM_DEFAULT_VALUES,
-  EVENT_STEPS,
-  EVENT_STEPS_TO_TEXT,
-} from '../form.const';
+import { EVENT_FORM_DEFAULT_VALUES, EVENT_STEPS } from '../../form.const';
 import { EventWithRoutesViewEntity } from '@/types/event.type';
 import {
   getPhaseAndEnabledStatus,
   EventEnabledStatus,
   EventPhase,
 } from '@/utils/event.util';
-import { Provider as JotaiProvider, useSetAtom } from 'jotai';
-import { eventAtom } from '../store/eventAtom';
-import { FormProvider, useForm, UseFormGetValues } from 'react-hook-form';
-import { BIG_REGIONS_TO_SHORT_NAME } from '@/constants/regions';
-import { EventFormValues } from '../form.type';
+import { Provider as JotaiProvider } from 'jotai';
+import { FormProvider, useForm } from 'react-hook-form';
+import { EventFormValues } from '../../form.type';
 import { ShuttleRoutesViewEntity } from '@/types/shuttleRoute.type';
-import { dailyEventIdsWithRoutesAtom } from '../store/dailyEventIdsWithRoutesAtom';
 import DemandCompleteScreen, {
   DemandCompleteStatus,
-} from './demand-complete-screen/DemandCompleteScreen';
-import { getIsLoggedIn } from '@/utils/handleToken.util';
-import { getUserDemands } from '@/services/demand.service';
-import { userDemandsAtom } from '../store/userDemandsAtom';
-import ExtraHubsInRouteStep from './steps/extra-steps/ExtraHubsInRouteStep';
+} from '../demand-complete-screen/DemandCompleteScreen';
+import ExtraHubsInRouteStep from '../steps/extra-steps/ExtraHubsInRouteStep';
 import { useGetShuttleRoutesOfEventWithPagination } from '@/services/shuttleRoute.service';
 import Button from '@/components/buttons/button/Button';
-import { userAlertRequestsAtom } from '../store/userAlertRequestsAtom';
-import { getUserAlertRequests } from '@/services/alertRequest.service';
 import FeedbackScreen from '@/components/feedback/FeedbackScreen';
+import useBottomSheetText from './hooks/useBottomSheetText';
+import useInputSectionText from './hooks/useInputSectionText';
+import useHistory from './hooks/useHistory';
+import useEventInitialization from './hooks/useEventInitialization';
+import useAlertFeedBackScreen from './hooks/useAlertFeedBackScreen';
 
 interface Props {
   event: EventWithRoutesViewEntity;
@@ -58,7 +51,7 @@ const EventForm = ({ event }: Props) => {
   const { phase, enabledStatus } = getPhaseAndEnabledStatus(event);
   const isDisabled = enabledStatus === 'disabled';
 
-  const { data: routesPages } = useGetShuttleRoutesOfEventWithPagination(
+  const { data: shuttleRoutesPages } = useGetShuttleRoutesOfEventWithPagination(
     {
       eventId: event.eventId,
       status: 'OPEN',
@@ -67,13 +60,14 @@ const EventForm = ({ event }: Props) => {
       enabled: event.hasOpenRoute,
     },
   );
-  const routes = useMemo(
-    () => routesPages?.pages.flatMap((page) => page.shuttleRoutes) ?? [],
-    [routesPages],
+  const shuttleRoutes = useMemo(
+    () => shuttleRoutesPages?.pages.flatMap((page) => page.shuttleRoutes) ?? [],
+    [shuttleRoutesPages],
   );
-  const isRoutesLoading = event.hasOpenRoute && routes.length === 0;
+  const isShuttleRoutesLoading =
+    event.hasOpenRoute && shuttleRoutes.length === 0;
 
-  if (isRoutesLoading) {
+  if (isShuttleRoutesLoading) {
     return (
       <div className="fixed bottom-0 left-0 right-0 z-10 mx-auto flex max-w-500 gap-8 bg-basic-white px-16 pb-24 pt-8">
         <Button
@@ -90,9 +84,9 @@ const EventForm = ({ event }: Props) => {
   return (
     <section className={isDisabled ? '' : 'px-16 pb-24'}>
       <JotaiProvider>
-        <Form
+        <Content
           event={event}
-          routes={routes}
+          shuttleRoutes={shuttleRoutes}
           phase={phase}
           enabledStatus={enabledStatus}
         />
@@ -103,127 +97,69 @@ const EventForm = ({ event }: Props) => {
 
 export default EventForm;
 
-interface FormProps {
+interface ContentProps {
   event: EventWithRoutesViewEntity;
-  routes: ShuttleRoutesViewEntity[];
+  shuttleRoutes: ShuttleRoutesViewEntity[];
   phase: EventPhase;
   enabledStatus: EventEnabledStatus;
 }
 
-const Form = ({ event, routes, phase, enabledStatus }: FormProps) => {
-  // 행사, 노선, 수요조사 상태 관리
-  const isInitialized = useRef(false);
-  const setEvent = useSetAtom(eventAtom);
-  const setDailyEventIdWithRoutes = useSetAtom(dailyEventIdsWithRoutesAtom);
+const Content = ({
+  event,
+  shuttleRoutes,
+  phase,
+  enabledStatus,
+}: ContentProps) => {
+  const { updateUserDemands, updateUserAlertRequests } = useEventInitialization(
+    {
+      event,
+      shuttleRoutes,
+    },
+  );
 
-  const setUserDemands = useSetAtom(userDemandsAtom);
-  const updateUserDemands = async () => {
-    const isLoggedIn = getIsLoggedIn();
-    if (!isLoggedIn) {
-      return;
-    }
-    const userDemands = await getUserDemands({
-      eventId: event.eventId,
-      status: 'OPEN',
-    });
-    setUserDemands(userDemands.shuttleDemands);
-  };
-
-  const setUserAlertRequests = useSetAtom(userAlertRequestsAtom);
-  const updateUserAlertRequests = async () => {
-    const isLoggedIn = getIsLoggedIn();
-    if (!isLoggedIn) {
-      return;
-    }
-    const userAlertRequests = await getUserAlertRequests();
-    const filteredUserAlertRequests =
-      userAlertRequests.shuttleRouteAlertRequests.filter(
-        (alertRequest) => alertRequest.shuttleRoute.eventId === event.eventId,
-      );
-    setUserAlertRequests(filteredUserAlertRequests);
-  };
-
-  useEffect(() => {
-    if (isInitialized.current) {
-      return;
-    }
-    isInitialized.current = true;
-    setEvent(event);
-    setDailyEventIdWithRoutes(routes);
-    updateUserDemands();
-    updateUserAlertRequests();
-  }, []);
-
-  // 폼 상태 관리
   const methods = useForm<EventFormValues>({
     defaultValues: EVENT_FORM_DEFAULT_VALUES,
   });
 
-  // 퍼널
   const initialStep = EVENT_STEPS[0];
   const { Funnel, Step, setStep, stepName } = useFunnel(
     EVENT_STEPS,
     initialStep,
   );
 
-  // stack 구조로 바텀시트 이동 기록 관리
-  const [history, setHistory] = useState<(typeof EVENT_STEPS)[number][]>([]);
-  const isBackButtonVisible = history.length > 0;
-
-  const setHistoryAndStep = (nextStep: (typeof EVENT_STEPS)[number]) => {
-    const currStep = stepName;
-    setHistory((prev) => [currStep, ...prev]);
-    setStep(nextStep);
-  };
-
-  const handleBack = () => {
-    const step = history[0];
-    if (!step) {
-      return;
-    }
-    setStep(step);
-    setHistory((prev) => prev.slice(1));
-  };
+  const { isHistoryAvailable, setHistoryAndStep, handleBack, resetHistory } =
+    useHistory({
+      stepName,
+      setStep,
+    });
 
   const onBottomSheetClose = () => {
+    resetHistory();
     setStep(initialStep);
-    setHistory([]);
     methods.reset(EVENT_FORM_DEFAULT_VALUES);
   };
 
-  // 바텀시트
   const { bottomSheetRef, contentRef, openBottomSheet, closeBottomSheet } =
     useBottomSheet({
       onClose: onBottomSheetClose,
     });
 
-  // 텍스트
   const { title: bottomSheetTitle, description: bottomSheetDescription } =
-    useMemo(() => getBottomSheetText(stepName, methods.getValues), [stepName]);
+    useBottomSheetText(stepName);
 
   const isReservationOpen =
     phase === 'reservation' && enabledStatus === 'enabled';
   const { title: inputSectionTitle, description: inputSectionDescription } =
-    useMemo(() => getInputSectionText(isReservationOpen), [isReservationOpen]);
+    useInputSectionText(isReservationOpen);
 
-  // 수요조사 완료 화면
   const [demandCompleteStatus, setDemandCompleteStatus] =
     useState<DemandCompleteStatus | null>(null);
 
-  // 빈자리 알림 신청 피드백 화면
-  const [
+  const {
     isAlertRequestFeedbackScreenOpen,
-    setIsAlertRequestFeedbackScreenOpen,
-  ] = useState(false);
-  const openAlertRequestFeedbackScreen = () => {
-    closeBottomSheet();
-    setTimeout(() => {
-      setIsAlertRequestFeedbackScreenOpen(true);
-    }, 100);
-  };
-  const closeAlertRequestFeedbackScreen = () => {
-    setIsAlertRequestFeedbackScreenOpen(false);
-  };
+    openAlertRequestFeedbackScreen,
+    closeAlertRequestFeedbackScreen,
+  } = useAlertFeedBackScreen({ closeBottomSheet });
 
   return (
     <>
@@ -250,7 +186,7 @@ const Form = ({ event, routes, phase, enabledStatus }: FormProps) => {
           ref={bottomSheetRef}
           title={bottomSheetTitle}
           description={bottomSheetDescription}
-          showBackButton={isBackButtonVisible}
+          showBackButton={isHistoryAvailable}
           onBack={handleBack}
         >
           <FormProvider {...methods}>
@@ -386,55 +322,4 @@ const Form = ({ event, routes, phase, enabledStatus }: FormProps) => {
       )}
     </>
   );
-};
-
-const getInputSectionText = (isReservationOpen: boolean) => {
-  if (isReservationOpen) {
-    return {
-      title: '노선을 확인해 보세요',
-      description: '노선 내 정류장과 출발 시간을 확인해 보세요.',
-    };
-  }
-  return {
-    title: '아직 확정된 노선이 없어요',
-    description:
-      '지금은 수요조사 기간이에요. 요청이 많은 정류장을 모아 노선을 만들고 있어요.',
-  };
-};
-
-const getBottomSheetText = (
-  stepName: (typeof EVENT_STEPS)[number],
-  getValues: UseFormGetValues<EventFormValues>,
-) => {
-  const text = EVENT_STEPS_TO_TEXT[stepName];
-
-  let titleInput = '';
-  let descriptionInput = '';
-
-  switch (stepName) {
-    case '[기타] 시/도 정보':
-      const sido = getValues('sido');
-      titleInput = BIG_REGIONS_TO_SHORT_NAME[sido];
-      break;
-    case '[기타] 복수 노선':
-      const hubsWithInfoForDuplicates = getValues('hubsWithInfoForDuplicates');
-      titleInput = hubsWithInfoForDuplicates?.length.toString() ?? '';
-      descriptionInput = hubsWithInfoForDuplicates?.[0].name ?? '';
-      break;
-    case '[기타] 빈자리 알림':
-      const selectedRouteForSeatAlarm = getValues('selectedRouteForSeatAlarm');
-      descriptionInput = selectedRouteForSeatAlarm?.name ?? '';
-      break;
-  }
-
-  const title =
-    typeof text.title === 'function' ? text.title(titleInput) : text.title;
-  const description =
-    typeof text.description === 'function'
-      ? text.description(descriptionInput)
-      : text.description;
-  return {
-    title,
-    description,
-  };
 };
